Fixed bug in reading in Poisson population rate from restart file
Former-commit-id: 21df5ebc633536a75a32b4a59ed7ec7661dceaa1

+++ cpp_sources/NeuroPop.cpp
@@ -1034,7 +1034,7 @@ void NeuroPop::import_restart(H5File& file, int pop_ind, string out_filename) {
 	str = pop_n + "/poisson_pop/";
 	if (group_exist_HDF5(file, str)) {
 		poisson_pop.on = 1;
-		poisson_pop.rate = read_scalar_HDF5<unsigned int>(file, str + string("rate")); 
+		poisson_pop.rate = read_scalar_HDF5<double>(file, str + string("rate")); 
 		read_vector_HDF5(file,str+string("next_spike_time"),poisson_pop.next_spike_time);
 	}
